chore: considated interfaces that interact with by-comp assemblies
Removes ExportInterface and InheritanceInterface class
Adds a single ByComponentInterface class to interact with the model
in terms of inheritance
