+++ trestle/core/crm/bycomp_interface.py
@@ -11,28 +11,35 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Provide interface to ssp allowing queries and operations for exports statements."""
+"""Provide interface to by-component allowing queries and operations for exports/inheritance statements."""
 
+import copy
 import logging
 from typing import Dict, List, Tuple
 
 import trestle.oscal.ssp as ossp
 from trestle.common.err import TrestleError
-from trestle.common.list_utils import as_dict, as_list
+from trestle.common.list_utils import as_dict, as_list, none_if_empty
 
 logger = logging.getLogger(__name__)
 
 
-class ExportInterface:
+class ByComponentInterface:
     """
-    Interface to query exported provided and responsibility statements from.
+    Interface to query and modify by-component assembly inheritance contents.
 
-    The by-component export statement is parsed and the responsibility and provided statements
+    The by-component is contained in two separate forms:  As an actual OSCAL by-component assembly,
+    and and multiple dicts providing direct lookup of inheritance statement by uuid.
+
+    The dicts are created by the ByComponentInterface constructor, parsed and the responsibility and provided statements
     are separated into three catagories:
 
     isolated responsibilities - A responsibility with no provided statement
     isolated provided - A provided statement with no referring responsibility statements
     export set - A set with a single responsibility and referred provided statement
+
+    For updating ByComponent inheritance and satisfied statements, the interface provides a method to reconcile the
+    the by-component assembly and merge input inherited and satisfied statements.
     """
 
     def __init__(self, by_comp: ossp.ByComponent):
@@ -42,6 +49,9 @@ class ExportInterface:
         self._provided_dict: Dict[str, ossp.Provided] = {}
         self._responsibility_dict: Dict[str, ossp.Responsibility] = {}
         self._responsibility_by_provided: Dict[str, List[ossp.Responsibility]] = {}
+
+        self._inherited_dict: Dict[str, ossp.Inherited] = self._create_inherited_dict()
+        self._satisfied_dict: Dict[str, ossp.Satisfied] = self._create_satisfied_dict()
 
         if by_comp.export:
             self._provided_dict = self._create_provided_dict()
@@ -72,6 +82,18 @@ class ExportInterface:
                 existing_list.append(responsibility)
                 responsibility_by_provided[responsibility.provided_uuid] = existing_list
         return responsibility_by_provided
+
+    def _create_inherited_dict(self) -> Dict[str, ossp.Inherited]:
+        inherited_dict: Dict[str, ossp.Inherited] = {}
+        for inherited in as_list(self._by_comp.inherited):
+            inherited_dict[str(inherited.provided_uuid)] = inherited
+        return inherited_dict
+
+    def _create_satisfied_dict(self) -> Dict[str, ossp.Satisfied]:
+        satisfied_dict: Dict[str, ossp.Satisfied] = {}
+        for satisfied in as_list(self._by_comp.satisfied):
+            satisfied_dict[str(satisfied.responsibility_uuid)] = satisfied
+        return satisfied_dict
 
     def get_isolated_responsibilities(self) -> List[ossp.Responsibility]:
         """Return all isolated exported responsibilities."""
@@ -105,6 +127,47 @@ class ExportInterface:
                 shared_responsibility: Tuple[ossp.Responsibility, ossp.Provided] = (responsibility, provided)
                 all_export_sets.append(shared_responsibility)
         return all_export_sets
+
+    def reconcile_inheritance_by_component(
+        self, incoming_inherited: List[ossp.Inherited], incoming_satisfied: List[ossp.Satisfied]
+    ) -> ossp.ByComponent:
+        """
+        Reconcile the inherited and satisfied statements in the by-component assembly with changes from the export.
+
+        Notes:
+            A statement is determined as existing if the provided uuid or responsibility uuid is in the existing in the
+            by-component assembly. If existing, the description will be updated if it has changed.
+
+            Any existing inherited or satisfied statements that are not in the incoming export will be removed.
+            If a statement is in the incoming export, but not in the existing by-component assembly, it will be added.
+        """
+        new_inherited: List[ossp.Inherited] = []
+        new_satisfied: List[ossp.Satisfied] = []
+
+        # Create a copy of the input by-component assembly to reconcile and return
+        new_by_comp: ossp.ByComponent = copy.deepcopy(self._by_comp)
+
+        for statement in incoming_inherited:
+            if statement.provided_uuid in self._inherited_dict:
+                existing_statement = self._inherited_dict[str(statement.provided_uuid)]
+                # Update the description if it has changed
+                existing_statement.description = statement.description
+                statement = existing_statement
+            new_inherited.append(statement)
+
+        new_by_comp.inherited = none_if_empty(new_inherited)
+
+        for statement in incoming_satisfied:
+            if statement.responsibility_uuid in self._satisfied_dict:
+                existing_statement = self._satisfied_dict[str(statement.responsibility_uuid)]
+                # Update the description if it has changed
+                existing_statement.description = statement.description
+                statement = existing_statement
+            new_satisfied.append(statement)
+
+        new_by_comp.satisfied = none_if_empty(new_satisfied)
+
+        return new_by_comp
 
     def _provided_has_responsibilities(self, provided_uuid: str) -> bool:
         """Return whether a provided UUID has responsibilities."""
